pyproject.toml
[project]
name = "VirtualTruck"
version = "0.2.0"
description = "Virtual Truck REST API with Django and Django Rest Framework"
authors = [
    {name = "Monitor Team", email = "monitor@monitor.com.br"},
]
dependencies = [
    "django>=5.1.3",
    "djangorestframework>=3.15.2",
    "django-simple-history>=3.7.0",
    "django-safedelete>=1.4.0",
    "drf-spectacular>=0.27.2",
    "setuptools>=75.1.0",
    "cpf-cnpj-validate>=1.4",
    "Pillow>=11.0.0",
    "python-magic>=0.4.27; sys_platform == \"linux\"",
    "pytest>=8.3.3",
    "pylint-django>=2.6.1",
    "faker>=33.0.0",
    "djangorestframework-simplejwt>=5.3.1",
    "pytz>=2024.2",
    "celery>=5.4.0",
    "python-dotenv>=1.0.1",
    "django-cors-headers>=4.5.0",
    "pytest>=8.3.3",
    "pillow>=11.0.0",
    "libmagic>=1.0",
    "django-filter>=24.3",
    "unidecode>=1.3.8",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "dj-database-url>=2.3.0",
    "orjson>=3.10.11",
    "baker>=1.3",
]

requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.isort]
profile = "black"

[tool.pdm.scripts]
manage = "python src/manage.py"
makemigrations = "pdm run manage makemigrations"
migrate = "pdm run manage migrate"
dev = "pdm run manage runserver 0.0.0.0:8000"
test = "pytest --ignore __pypackages__"
populate = "python src/manage.py populate"
celery = "celery -A django_project worker -l info"


[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-django>=4.9.0",
    "colorama>=0.4.6",
    "model-bakery>=1.20.0",
    "pylint>=3.3.1",
    "isort>=5.13.2",
]
